+++ internal/game/ambient_test.go
@@ -33,7 +33,7 @@ func arenePeuplee(t *testing.T, figurants int) *World {
 	}
 
 	g := NewCostGrid(32, 32)
-	w := NewWorld(profils, armes, progressionLivree(t), caissesLivrees(t), sansVagues(), g,
+	w := NewWorld(profils, armes, progressionLivree(t), caissesLivrees(t), fiolesLivrees(t), sansVagues(), g,
 		graineDeTest, capacitesDeTest)
 	w.Place(FromInt(16)+One/2, FromInt(16)+One/2)
 	for range figurants {